requirements.txt
Pillow==9.3.0
py_console==0.1.4
py_cpuinfo==9.0.0
setuptools==65.6.3
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
pipdeptree==2.2.0
